Blueprints: add blueprint's images empty state
